+++ Editor/CityImport/CityLoadConfigGUI.cs
@@ -2,10 +2,8 @@
 using PLATEAU.CityImport.Config;
 using PLATEAU.CityImport.Config.PackageLoadConfigs;
 using PLATEAU.Editor.CityImport.PackageLoadConfigGUIs;
-using PLATEAU.Editor.CityImport.PackageLoadConfigGUIs.GUIParts;
+using PLATEAU.Editor.CityImport.PackageLoadConfigGUIs.Components;
 using PLATEAU.Editor.EditorWindow.Common;
-using UnityEditor;
-using UnityEngine;
 
 namespace PLATEAU.Editor.CityImport
 {
