+++ Editor/CityImport/PackageLoadConfigGUIs/Extendables/Components/SetAttrInfoGUI.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 5d087e4abeef4fada317aee791623a50
+timeCreated: 1698639905
