+++ Editor/CityImport/PackageLoadConfigGUIs/Extendables/ExtendableConfigBase.cs
@@ -1,0 +1,20 @@
+using PLATEAU.CityImport.Config.PackageLoadConfigs;
+
+namespace PLATEAU.Editor.CityImport.PackageLoadConfigGUIs.Extendables
+{
+    /// <summary>
+    /// 一括設定の中にある各項目を抽象化したものです。
+    /// <see cref="Draw"/>で設定GUIを描画します。
+    /// </summary>
+    internal abstract class ExtendableConfigGUIBase
+    {
+        protected PackageLoadConfigExtendable Conf;
+
+        public ExtendableConfigGUIBase(PackageLoadConfigExtendable conf)
+        {
+            Conf = conf;
+        }
+        
+        public abstract void Draw();
+    }
+}
