+++ Editor/CityImport/PackageLoadConfigGUIs/Extendables/ExtendableConfigBase.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 8d14f0bf81e94d008f2b1d8fe19de890
+timeCreated: 1698639676
