+++ Editor/CityImport/PackageLoadConfigGUIs/PackageLoadConfigGUI.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
-using PLATEAU.CityImport.Config;
 using PLATEAU.CityImport.Config.PackageLoadConfigs;
 using PLATEAU.Dataset;
-using PLATEAU.Editor.CityImport.PackageLoadConfigGUIs.GUIParts;
+using PLATEAU.Editor.CityImport.PackageLoadConfigGUIs.Components;
+using PLATEAU.Editor.CityImport.PackageLoadConfigGUIs.Extendables;
 using PLATEAU.Editor.EditorWindow.Common;
 using UnityEditor;
