+++ Editor/EditorWindow/PlateauWindow/MainTabGUI/ImportGUIParts/ImportButton.cs
@@ -4,7 +4,6 @@ using PLATEAU.CityImport.Config;
 using PLATEAU.Editor.EditorWindow.Common;
 using PLATEAU.Util;
 using PLATEAU.Util.Async;
-using UnityEditor;
 
 namespace PLATEAU.Editor.EditorWindow.PlateauWindow.MainTabGUI.ImportGUIParts
 {
@@ -57,7 +56,7 @@ namespace PLATEAU.Editor.EditorWindow.PlateauWindow.MainTabGUI.ImportGUIParts
                         bool dialogueResult = Dialogue.Display($"インポートをキャンセルしますか？", "はい", "いいえ");
                         if (dialogueResult)
                         {
-                            cancellationTokenSrc.Cancel();
+                            cancellationTokenSrc?.Cancel();
                         }
                     }
                 }
